final
the game work. a few changes are needed though. If the snake's body on the apple it erase it. It could be handled with adding the color for the queue of pixles.  Unfortunately, I dont have the time for those changes (considare the fact that it was merely self learning side project of mine, on my free time while Im a student).